services/ui-service/requirements.txt
streamlit==1.31.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.6.1
asyncio==3.4.3
websockets==12.0
